requirements.txt
numpy
opencv-python
Pillow
google-cloud-vision
tk
